# packages/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "properties": {
          "ip_address": {
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "type": "object"
      },
      "CreateSessionResponse": {
        "properties": {
          "session": {
            "properties": {
              "created_at": {
                "type": "string"
              },
              "expires_at": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "ip_address": {
                "type": "string"
              },
              "org_id": {
                "type": "string"
              },
              "revoked_at": {
                "type": "string"
              },
              "user_agent": {
                "type": "string"
              },
              "user_id": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "session",
          "token"
        ],
        "type": "object"
      },
...
      "ListResponse": {
        "properties": {
          "items": {},
          "next_cursor": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "revoked_at": {
            "format": "date-time",
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "org_id",
          "created_at",
          "expires_at"
        ],
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "PAT or session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Identity and access management API. Schema-driven, extensible, open source.",
    "title": "Zitadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            },
            "description": "List sessions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List sessions",
        "tags": [
          "Sessions"
        ]
      },
      "post": {
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Create a session"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a session",
        "tags": [
          "Sessions"
        ]
      }
    },
    "/v1/sessions/{id}": {
      "get": {
        "operationId": "getSession",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Get a session"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a session",
        "tags": [
          "Sessions"
        ]
      }
    },
    "/v1/sessions/{id}/revoke": {
      "post": {
        "operationId": "revokeSession",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Revoke a session"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke a session",
        "tags": [
          "Sessions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This instance",
      "url": "/"
    }
  ],
...
}